Move store and mock store into separate folder
