+++ user-registry-ui/store/index.js
@@ -1,0 +1,1 @@
+export { createStore } from './store'
