+++ user-registry-ui/store/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from 'redux-starter-kit'
-import { githubReducer } from './github'
-import { ethereumReducer } from './ethereum'
+import { githubReducer } from '../github'
+import { ethereumReducer } from '../ethereum'
 
 export const createStore = () => configureStore({
   reducer: {
